refactor: restructure package hierarcy by moving nested packages out
